# api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: Terminarz
# during dev, should point to your local machine
host: localhost:10010
# basePath prefixes all resource paths 
basePath: /
# 
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
paths:
  /events:
      x-swagger-router-controller: event
      get:
        operationId: readEvents
        responses:
          "200":
            description: Success
            schema:
              $ref: "#/definitions/EventListResponse"
          default:
            description: Error
            schema:
              $ref: "#/definitions/ErrorResponse"
      post:
        operationId: createEvent
        parameters:
          - name: event
            in: body
            required: true
            schema:
              $ref: "#/definitions/Event"
        responses:
          "200":
            description: Success
            schema:
              $ref: "#/definitions/EventResponse"
          default:
            description: Error
            schema:
              $ref: "#/definitions/ErrorResponse"
  /events/{id}:
      x-swagger-router-controller: event
      get:
        operationId: readEvent
        parameters:
          - name: id
            type: string
            in: path
            required: true
        responses:
          "200":
            description: Success
            schema:
              $ref: "#/definitions/EventResponse"
          default:
            description: Error
            schema:
              $ref: "#/definitions/ErrorResponse"
      put:
        operationId: updateEvent
        parameters:
          - name: id
            type: string
            in: path
            required: true
          - name: event
            in: body
            required: true
            schema:
              $ref: "#/definitions/Event"
        responses:
          "200":
            description: Success
            schema:
              $ref: "#/definitions/EventResponse"
          default:
            description: Error
            schema:
              $ref: "#/definitions/ErrorResponse"
      delete:
        operationId: deleteEvent
        parameters:
          - name: id
            type: string
            in: path
            required: true
        responses:
          "200":
            description: Success
            schema:
              $ref: "#/definitions/EventResponse"
          default:
            description: Error
            schema:
              $ref: "#/definitions/ErrorResponse"
  /users:
      x-swagger-router-controller: user
      post:
        operationId: registerUser
        parameters:
          - name: user
            in: body
            required: true
            schema:
              $ref: "#/definitions/User"
        responses:
          "200":
            description: Success
            schema:
              $ref: "#/definitions/SuccessResponse"
          default:
            description: Error
            schema:
              $ref: "#/definitions/ErrorResponse"
  /users/{id}:
      x-swagger-router-controller: user
      get:
        operationId: readUser
        parameters:
          - name: id
            type: string
            in: path
            required: true
        responses:
          "200":
            description: Success
            schema:
              $ref: "#/definitions/UserResponse"
          default:
            description: Error
            schema:
              $ref: "#/definitions/ErrorResponse"
      put:
        operationId: updateUser
        parameters:
          - name: id
            type: string
            in: path
            required: true
          - name: user
            in: body
            required: true
            schema:
              $ref: "#/definitions/UserUpdate"
        responses:
          "200":
            description: Success
            schema:
              $ref: "#/definitions/UserResponse"
          default:
            description: Error
            schema:
              $ref: "#/definitions/ErrorResponse"
      delete:
        operationId: deleteUser
        parameters:
          - name: id
            type: string
            in: path
            required: true
        responses:
          "200":
            description: Success
            schema:
              $ref: "#/definitions/UserResponse"
          default:
            description: Error
            schema:
              $ref: "#/definitions/ErrorResponse"
  /swagger:
    x-swagger-pipe: swagger_raw
# complex objects have schema definitions
definitions:
  EventResponse:
    type: object
    properties:
      id:
        type: string
      name:
        type: string
      description:
        type: string
      startDateTime:
        type: string
        format: date-time
      endDateTime:
        type: string
        format: date-time
      public:
        type: boolean
  EventListResponse:
    properties:
      events:
        type: array
        items:
          $ref: "#/definitions/EventResponse"
  Event:
    type: object
    properties:
      name:
        type: string
      description:
        type: string
      startDateTime:
        type: string
        format: date-time
      endDateTime:
        type: string
        format: date-time
      public:
        type: boolean
  UserResponse:
    type: object
    properties:
      name:
        type: string
  User:
    type: object
    properties:
      name:
        type: string
      password:
        type: string
  UserUpdate:
    type: object
    properties:
      password:
        type: string

  ErrorResponse:
    properties:
      message:
        type: string
  SuccessResponse:
    properties:
      message:
        type: string